% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{check_ap}
\alias{check_ap}
\title{Checks whether antipsychotic names are in the key}
\usage{
check_ap(x, key = chlorpromazineR::gardner2010, ap_label, route,
  route_label)
}
\arguments{
\item{x}{data.frame with antipsychotic name and dose data}

\item{key}{source of the conversion factors--defaults to Gardner et al. 2010}

\item{ap_label}{column in x that stores antipsychotic name}

\item{route}{options include "oral", "sai", "lai" or "mixed"}

\item{route_label}{if "mixed" route is specified, provide the column that
stores the route information}
}
\value{
number of antipsychotic names in x[,ap_label] that don't match key
}
\description{
Provided a data.frame, x, this checks that the antipsychotic names stored in
the x's variable ap_label are present in the key.
}
\concept{checking functions}
