% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_functions.R
\name{check_key}
\alias{check_key}
\title{Check whether a conversion key is the expected format}
\usage{
check_key(key)
}
\arguments{
\item{key}{the key to check}
}
\value{
TRUE if the key is valid, otherwise a error is thrown.
}
\description{
chlorpromazineR uses conversion factors stored in a named list of 3 named
lists. This verifies that the key is in a usable format, which can be helpful
when creating custom keys or modifying included keys.
}
\examples{
check_key(gardner2010)
}
\seealso{
Other key functions: \code{\link{add_key}},
  \code{\link{trim_key}}
}
\concept{key functions}
