% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{to_cpz}
\alias{to_cpz}
\title{Calculates chlorpromazine-equivalent doses}
\usage{
to_cpz(x, ap_label, dose_label, route = "oral",
  key = chlorpromazineR::gardner2010, eq_label = "cpz_eq",
  factor_label = "cpz_conv_factor", route_label = NULL, q = NULL)
}
\arguments{
\item{x}{data.frame with antipsychotic name and dose data}

\item{ap_label}{column in x that stores antipsychotic name}

\item{dose_label}{column in x that stores dose}

\item{route}{options include "oral", "sai", "lai" or "mixed"}

\item{key}{source of the conversion factors--defaults to Gardner et al. 2010}

\item{eq_label}{the name of the column to be created, to save the 
calculated CPZ-equivalent dose}

\item{factor_label}{the name of the column to be created to store the
conversion factors}

\item{route_label}{if "mixed" route is specified, provide the column that
stores the route information}

\item{q}{if long-acting injectable doses are included, provide the column
that stores the injection frequency (days)}
}
\value{
data.frame with new variables storing conversion factor and 
CPZ-equivalent doses
}
\description{
If the time_unit and activity_unit attributes are already in the data.frame,
they do not need to be set again, but otherwise they will need to be
specified in the input parameters.
}
\concept{conversion functions}
